requirements.txt
django
tensorflow
pillow
numpy
requests
gunicorn
google-generativeai
python-dotenv
